+++ src/Abstracts/UnionType.php
@@ -18,13 +18,6 @@ if ( ! class_exists( '\AxeWP\GraphQL\Abstracts\UnionType' ) ) {
 		use TypeResolverTrait;
 
 		/**
-		 * The WPGraphQL TypeRegistry instance.
-		 *
-		 * @var ?\WPGraphQL\Registry\TypeRegistry
-		 */
-		protected static $type_registry;
-
-		/**
 		 * Gets the array of possible GraphQL types that can be resolved to.
 		 *
 		 * @return string[]
@@ -33,12 +26,8 @@ if ( ! class_exists( '\AxeWP\GraphQL\Abstracts\UnionType' ) ) {
 
 		/**
 		 * {@inheritDoc}
-		 *
-		 * @param \WPGraphQL\Registry\TypeRegistry|null $type_registry The WPGraphQL TypeRegistry instance.
 		 */
-		public static function register( $type_registry = null ): void {
-			self::$type_registry = $type_registry;
-
+		public static function register(): void {
 			register_graphql_union_type( static::get_type_name(), static::get_type_config() );
 		}