+++ src/Traits/TypeResolverTrait.php
@@ -9,35 +9,22 @@ namespace AxeWP\GraphQL\Traits;
 
 use Closure;
 use Error;
+use WPGraphQL;
 
 if ( ! trait_exists( '\AxeWP\GraphQL\Traits\TypeResolverTrait' ) ) {
 
 	/**
 	 * Trait - TypeResolverTrait
-	 *
-	 * @property ?\WPGraphQL\Registry\TypeRegistry $type_registry The WPGraphQL TypeRegistry instance.
 	 */
 	trait TypeResolverTrait {
 		/**
 		 * The function used to resolve the Interface type in the `resolveType` callback.
-		 *
-		 * @throws \Error If $type_registry is not set.
 		 */
 		protected static function get_type_resolver(): Closure {
 			/**
 			 * @param mixed       $value The value from the resolver of the parent field.
 			 */
 			return static function ( $value ) {
-				if ( ! static::$type_registry instanceof \WPGraphQL\Registry\TypeRegistry ) {
-					throw new Error(
-						sprintf(
-						// translators: function name.
-							__( 'Incorrect usage of %s. This method may only be called after self::$type_registry is set.', 'wp-graphql-plugin-name' ),
-							__FUNCTION__
-						)
-					);
-				}
-
 				$type_name = static::get_resolved_type_name( $value );
 
 				if ( empty( $type_name ) ) {
@@ -47,7 +34,9 @@ if ( ! trait_exists( '\AxeWP\GraphQL\Traits\TypeResolverTrait' ) ) {
 					);
 				}
 
-				return static::$type_registry->get_type( $type_name );
+				$type_registry = WPGraphQL::get_type_registry();
+
+				return $type_registry->get_type( $type_name );
 			};
 		}